backend/requirements.txt
boto3==1.28.85
botocore==1.31.85
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
PyJWT==2.8.0
bcrypt==4.1.1
